pyproject.toml
[tool.poetry]
name = "coursepaper-5"
version = "0.1.0"
description = ""
authors = ["Zhorik Zeniuk <ZhorikZeniuk@gmail.ru>"]
readme = "README.md"
packages = [{include = "coursepaper_5"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
